requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
tldextract==5.3.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2.0
gunicorn==21.2.0
supabase==1.0.3
python-dotenv==1.0.1
